pyproject.toml
[tool.poetry]
name = "django-simple-slack-app"
version = "0.3.0"
description = "nice toolkit for Slack app(bot) that support Event API & OAuth"
authors = ["Matthew, Lee <bluedisk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.4"
slackclient = "^2.5.0"
slackeventsapi = "^2.1.0"
hgtk = "^0.1.3"
pyee = "^7.0.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
